ai-crew/requirements.txt
crewai
crewai-tools
langchain-google-genai
langchain-openai
python-dotenv
requests
